Move box resolvers into dedicated modules
