+++ back/boxtribute_server/graph_ql/resolvers.py
@@ -1,18 +1,15 @@
 """GraphQL resolver functionality"""
 from ariadne import MutationType, QueryType, convert_kwargs_to_snake_case
-from flask import g
 
 from ..authz import authorize, authorized_bases_filter
 from ..enums import LocationType
-from ..models.crud import create_box, create_qr_code, get_box_history, update_box
+from ..models.crud import create_qr_code
 from ..models.definitions.box import Box
 from ..models.definitions.location import Location
 from ..models.definitions.product import Product
 from ..models.definitions.qr_code import QrCode
-from ..models.definitions.tag import Tag
 from .bindables import (
     beneficiary,
-    box,
     classic_location,
     product,
     qr_code,
@@ -37,22 +34,10 @@ def resolve_qr_exists(*_, qr_code):
 
 
 @query.field("qrCode")
-@box.field("qrCode")
 @convert_kwargs_to_snake_case
 def resolve_qr_code(obj, _, qr_code=None):
     authorize(permission="qr:read")
     return obj.qr_code if qr_code is None else QrCode.get(QrCode.code == qr_code)
-
-
-@box.field("tags")
-def resolve_box_tags(box_obj, info):
-    return info.context["tags_for_box_loader"].load(box_obj.id)
-
-
-@box.field("history")
-def resolve_box_history(box_obj, _):
-    authorize(permission="history:read")
-    return get_box_history(box_obj.id)
 
 
 @query.field("product")
@@ -62,28 +47,9 @@ def resolve_product(*_, id):
     return product
 
 
-@box.field("product")
 @unboxed_items_collection.field("product")
 def resolve_box_product(obj, info):
     return info.context["product_loader"].load(obj.product_id)
-
-
-@box.field("size")
-def resolve_size(box_obj, info):
-    return info.context["size_loader"].load(box_obj.size_id)
-
-
-@query.field("box")
-@convert_kwargs_to_snake_case
-def resolve_box(*_, label_identifier):
-    box = (
-        Box.select(Box, Location)
-        .join(Location)
-        .where(Box.label_identifier == label_identifier)
-        .get()
-    )
-    authorize(permission="stock:read", base_id=box.location.base_id)
-    return box
 
 
 @query.field("location")
@@ -92,12 +58,6 @@ def resolve_location(obj, _, id):
     if location.type == LocationType.ClassicLocation:
         authorize(permission="location:read", base_id=location.base_id)
         return location
-
-
-@box.field("location")
-def resolve_box_location(obj, _):
-    authorize(permission="location:read", base_id=obj.location.base_id)
-    return obj.location
 
 
 @query.field("locations")
@@ -117,12 +77,6 @@ def resolve_products(*_, pagination_input=None):
     )
 
 
-@box.field("state")
-def resolve_box_state(box_obj, _):
-    # Instead of a BoxState instance return an integer for EnumType conversion
-    return box_obj.state_id
-
-
 @classic_location.field("defaultBoxState")
 def resolve_location_default_box_state(location_obj, _):
     # Instead of a BoxState instance return an integer for EnumType conversion
@@ -134,50 +88,6 @@ def resolve_location_default_box_state(location_obj, _):
 def resolve_create_qr_code(*_, box_label_identifier=None):
     authorize(permission="qr:create")
     return create_qr_code(box_label_identifier=box_label_identifier)
-
-
-@mutation.field("createBox")
-@convert_kwargs_to_snake_case
-def resolve_create_box(*_, creation_input):
-    requested_location = Location.get_by_id(creation_input["location_id"])
-    authorize(permission="stock:write", base_id=requested_location.base_id)
-    authorize(permission="location:read", base_id=requested_location.base_id)
-    requested_product = Product.get_by_id(creation_input["product_id"])
-    authorize(permission="product:read", base_id=requested_product.base_id)
-
-    tag_ids = creation_input.get("tag_ids", [])
-    for t in Tag.select().where(Tag.id << tag_ids):
-        authorize(permission="tag_relation:assign", base_id=t.base_id)
-
-    return create_box(user_id=g.user.id, **creation_input)
-
-
-@mutation.field("updateBox")
-@convert_kwargs_to_snake_case
-def resolve_update_box(*_, update_input):
-    box = (
-        Box.select(Box, Location)
-        .join(Location)
-        .where(Box.label_identifier == update_input["label_identifier"])
-        .get()
-    )
-    authorize(permission="stock:write", base_id=box.location.base_id)
-
-    location_id = update_input.get("location_id")
-    if location_id is not None:
-        requested_location = Location.get_by_id(location_id)
-        authorize(permission="location:read", base_id=requested_location.base_id)
-
-    product_id = update_input.get("product_id")
-    if product_id is not None:
-        requested_product = Product.get_by_id(product_id)
-        authorize(permission="product:read", base_id=requested_product.base_id)
-
-    tag_ids = update_input.get("tag_ids", [])
-    for t in Tag.select().where(Tag.id << tag_ids):
-        authorize(permission="tag_relation:assign", base_id=t.base_id)
-
-    return update_box(user_id=g.user.id, **update_input)
 
 
 @classic_location.field("boxes")
